+++ tests/testthat/test-corr_hm_app.R
@@ -25,7 +25,7 @@ ID <- poc( # nolint
   )
 )
 
-root_app <- start_app_driver(dv.biomarker.general::mock_app_corr_hm())
+root_app <- start_app_driver(dv.explorer.parameter::mock_app_corr_hm())
 on.exit(if ("stop" %in% names(root_app)) root_app$stop())
 
 fail_if_app_not_started <- function() {
@@ -157,7 +157,7 @@ test_that("default values are set", {
 
   app <- start_app_driver(
     rlang::quo(
-      dv.biomarker.general::mock_app_corr_hm(
+      dv.explorer.parameter::mock_app_corr_hm(
         ui_defaults = !!ui_defaults,
         srv_defaults = !!srv_defaults
       )